-Modifications to allow compilation using platformIO -- Formatting mods -- Some modifications to file formats -- Changes filenames from .ino to .cpp -- Added Arduino.h header files where necessary -- Alarm code still a work in progress
