+++ blower.cpp
@@ -15,27 +15,10 @@
   You should have received a copy of the GNU General Public License
   along with FixMoreLungs.  If not, see <https://www.gnu.org/licenses/>.
 */
+#include <Arduino.h>
 
-#include "ventilator.h"
-#include "comms.h"
+#include "blower.h"
 
-void ventilator_start() {
+void blower_init() {
 
-    alarm_init();
-
-    comms_sendResetState(); // Inform the Interface Controller that we just started/restarted
-
-	for (;;) {	
-		ventilator_control();
-	}
-}
-
-static void ventilator_control() {
-
-	// Check any new commands?
-    comms_handler();
-
-    // pid_execute();
-    //comms_sendResetState(); 
-    watchdog_handler();
 }
