+++ pid.cpp
@@ -15,6 +15,8 @@
   You should have received a copy of the GNU General Public License
   along with FixMoreLungs.  If not, see <https://www.gnu.org/licenses/>.
 */
+#include <Arduino.h>
+
 #include "pid.h"
 
 //Define Variables we'll be connecting to
