+++ sensors.cpp
@@ -15,10 +15,10 @@
   You should have received a copy of the GNU General Public License
   along with FixMoreLungs.  If not, see <https://www.gnu.org/licenses/>.
 */
+#include <Arduino.h>
 
-#include "version.h"
+#include "sensors.h"
 
+void sensors_init() {
 
-char * version_getVersion() {
-	return (char *) SW_VERSION;
 }
